Chnage structure of program
1) rename "snake.cpp" to "main.cpp".
2) add "functions.cpp".
3) Add more comments.
4) Add description in README file.
