+++ Makefile
@@ -1,7 +1,7 @@
 CXX = g++
 CXXFLAGS = -Wall -std=c++17 -lncurses
 DEPS = snake.h
-OBJ = snake.o class_snake.o
+OBJ = functions.o class_snake.o main.o
 
 # Default goal
 snake: $(OBJ)
@@ -13,4 +13,4 @@ $(OBJ): $(DEPS)
 
 .PHONY: clean
 clean :
-	rm snake
+	rm -f *.o snake
