+++ main.cpp
@@ -6,8 +6,7 @@
 
 WINDOW *win = NULL;
 unsigned int speed = 100;
-/* ------------------------------------------------------------------------- */
-void create_border(void);
+
 /* ------------------------------------------------------------------------- */
 
 /* Handle the program if it is terminated using kill(), pkill() or 'ctrl+C' */
@@ -144,29 +143,6 @@ int main(){
   return 0;
 }
 
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-
-void create_border(void){
-  if(wborder(win, 0, 0, 0, 0, 0, 0, 0, 0) != OK){
-    fprintf(stderr,"\nError wborder(): %d\n", ERR);
-    delwin(win);
-    echo();
-    endwin();
-    exit(0);
-  }
-}
-
-/* ------------------------------------------------------------------------- */
-
-
-/* ------------------------------------------------------------------------- */
-
-
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
 /* ------------------------------------------------------------------------- */
 /* ------------------------------------------------------------------------- */
 /* ------------------------------------------------------------------------- */
